split connection, enable nullable and remove SoftProtocolException
