+++ projects/RabbitMQ.Client/client/impl/Connection.cs
@@ -32,10 +32,7 @@
 using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Net.Sockets;
-using System.Reflection;
 using System.Runtime.CompilerServices;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -43,45 +40,26 @@ using RabbitMQ.Client.Events;
 using RabbitMQ.Client.Exceptions;
 using RabbitMQ.Client.Impl;
 using RabbitMQ.Client.Logging;
-using RabbitMQ.Util;
 
 namespace RabbitMQ.Client.Framing.Impl
 {
-    internal sealed class Connection : IConnection
+    #nullable enable
+    internal sealed partial class Connection : IConnection
     {
-        private static readonly string s_version = typeof(Connection).Assembly
-                                            .GetCustomAttribute<AssemblyInformationalVersionAttribute>()
-                                            .InformationalVersion;
-
         private bool _disposed;
         private volatile bool _closed;
 
         private readonly IConnectionFactory _factory;
-        private readonly IFrameHandler _frameHandler;
         private readonly ModelBase _model0;
         private readonly MainSession _session0;
 
         private Guid _id = Guid.NewGuid();
-        private volatile bool _running = true;
         private SessionManager _sessionManager;
 
-        //
-        // Heartbeats
-        //
-        private TimeSpan _heartbeat = TimeSpan.Zero;
-        private TimeSpan _heartbeatTimeSpan = TimeSpan.FromSeconds(0);
-        private int _missedHeartbeats;
-        private bool _heartbeatDetected;
+        private ShutdownEventArgs? _closeReason;
+        public ShutdownEventArgs? CloseReason => Volatile.Read(ref _closeReason);
 
-        private Timer _heartbeatWriteTimer;
-        private Timer _heartbeatReadTimer;
-
-        private Task _mainLoopTask;
-
-        private ShutdownEventArgs _closeReason;
-        public ShutdownEventArgs CloseReason => Volatile.Read(ref _closeReason);
-
-        public Connection(IConnectionFactory factory, IFrameHandler frameHandler, string clientProvidedName = null)
+        public Connection(IConnectionFactory factory, IFrameHandler frameHandler, string? clientProvidedName = null)
         {
             ClientProvidedName = clientProvidedName;
             _factory = factory;
@@ -94,14 +72,48 @@ namespace RabbitMQ.Client.Framing.Impl
             _connectionShutdownWrapper = new EventingWrapper<ShutdownEventArgs>("OnShutdown", onException);
 
             _sessionManager = new SessionManager(this, 0);
-            _session0 = new MainSession(this) { Handler = NotifyReceivedCloseOk };
+            _session0 = new MainSession(this);
             _model0 = (ModelBase)Protocol.CreateModel(factory, _session0);
 
-            StartMainLoop();
+            ClientProperties = new Dictionary<string, object?>(factory.ClientProperties)
+            {
+                ["capabilities"] = Protocol.Capabilities,
+                ["connection_name"] = ClientProvidedName
+            };
+
+            _mainLoopTask = Task.Factory.StartNew(MainLoop, TaskCreationOptions.LongRunning);
             Open();
         }
 
         public Guid Id => _id;
+
+        public string? ClientProvidedName { get; }
+
+        public ushort ChannelMax => _sessionManager.ChannelMax;
+
+        public IDictionary<string, object?> ClientProperties { get; private set; }
+
+        public AmqpTcpEndpoint Endpoint => _frameHandler.Endpoint;
+
+        public uint FrameMax { get; private set; }
+
+        public bool IsOpen => CloseReason is null;
+
+        public int LocalPort => _frameHandler.LocalPort;
+
+        ///<summary>Another overload of a Protocol property, useful
+        ///for exposing a tighter type.</summary>
+        internal ProtocolBase Protocol => (ProtocolBase)Endpoint.Protocol;
+
+        public int RemotePort => _frameHandler.RemotePort;
+
+        public IDictionary<string, object?>? ServerProperties { get; private set; }
+
+        public IList<ShutdownReportEntry> ShutdownReport => _shutdownReport;
+        private ShutdownReportEntry[] _shutdownReport = Array.Empty<ShutdownReportEntry>();
+
+        ///<summary>Explicit implementation of IConnection.Protocol.</summary>
+        IProtocol IConnection.Protocol => Endpoint.Protocol;
 
         public event EventHandler<CallbackExceptionEventArgs> CallbackException
         {
@@ -129,13 +141,14 @@ namespace RabbitMQ.Client.Framing.Impl
             add
             {
                 ThrowIfDisposed();
-                if (IsOpen)
+                var reason = CloseReason;
+                if (reason is null)
                 {
                     _connectionShutdownWrapper.AddHandler(value);
                 }
                 else
                 {
-                    value(this, CloseReason);
+                    value(this, reason);
                 }
             }
             remove
@@ -182,58 +195,41 @@ namespace RabbitMQ.Client.Framing.Impl
             remove { }
         }
 
-        public string ClientProvidedName { get; }
-
-        public ushort ChannelMax => _sessionManager.ChannelMax;
-
-        public IDictionary<string, object> ClientProperties { get; private set; }
-
-        public AmqpTcpEndpoint Endpoint => _frameHandler.Endpoint;
-
-        public uint FrameMax { get; private set; }
-
-        public TimeSpan Heartbeat
+        internal void TakeOver(Connection other)
         {
-            get => _heartbeat;
-            set
+            _callbackExceptionWrapper.Takeover(other._callbackExceptionWrapper);
+            _connectionBlockedWrapper.Takeover(other._connectionBlockedWrapper);
+            _connectionUnblockedWrapper.Takeover(other._connectionUnblockedWrapper);
+            _connectionShutdownWrapper.Takeover(other._connectionShutdownWrapper);
+        }
+
+        public IModel CreateModel()
+        {
+            EnsureIsOpen();
+            ISession session = CreateSession();
+            var model = (ModelBase)Protocol.CreateModel(_factory, session);
+            model.ContinuationTimeout = _factory.ContinuationTimeout;
+            model._Private_ChannelOpen();
+            return model;
+        }
+
+        internal ISession CreateSession()
+        {
+            return _sessionManager.Create();
+        }
+
+        internal void EnsureIsOpen()
+        {
+            if (!IsOpen)
             {
-                _heartbeat = value;
-                // timers fire at slightly below half the interval to avoid race
-                // conditions
-                _heartbeatTimeSpan = TimeSpan.FromMilliseconds(_heartbeat.TotalMilliseconds / 4);
-                _frameHandler.ReadTimeout = TimeSpan.FromMilliseconds(_heartbeat.TotalMilliseconds * 2);
+                ThrowAlreadyClosedException(CloseReason!);
             }
         }
 
-        public bool IsOpen => CloseReason is null;
-
-        public int LocalPort => _frameHandler.LocalPort;
-
-        ///<summary>Another overload of a Protocol property, useful
-        ///for exposing a tighter type.</summary>
-        internal ProtocolBase Protocol => (ProtocolBase)Endpoint.Protocol;
-
-        public int RemotePort => _frameHandler.RemotePort;
-
-        public IDictionary<string, object> ServerProperties { get; private set; }
-
-        public IList<ShutdownReportEntry> ShutdownReport => _shutdownReport;
-        private ShutdownReportEntry[] _shutdownReport = Array.Empty<ShutdownReportEntry>();
-
-        ///<summary>Explicit implementation of IConnection.Protocol.</summary>
-        IProtocol IConnection.Protocol => Endpoint.Protocol;
-
-        internal static IDictionary<string, object> DefaultClientProperties()
+        ///<summary>API-side invocation of connection.close with timeout.</summary>
+        public void Close(ushort reasonCode, string reasonText, TimeSpan timeout, bool abort)
         {
-            var table = new Dictionary<string, object>(5)
-            {
-                ["product"] = Encoding.UTF8.GetBytes("RabbitMQ"),
-                ["version"] = Encoding.UTF8.GetBytes(s_version),
-                ["platform"] = Encoding.UTF8.GetBytes(".NET"),
-                ["copyright"] = Encoding.UTF8.GetBytes("Copyright (c) 2007-2020 VMware, Inc."),
-                ["information"] = Encoding.UTF8.GetBytes("Licensed under the MPL. See https://www.rabbitmq.com/")
-            };
-            return table;
+            Close(new ShutdownEventArgs(ShutdownInitiator.Application, reasonCode, reasonText), abort, timeout);
         }
 
         ///<summary>Try to close connection in a graceful way</summary>
@@ -257,19 +253,18 @@ namespace RabbitMQ.Client.Framing.Impl
             {
                 if (!abort)
                 {
-                    throw new AlreadyClosedException(CloseReason);
+                    ThrowAlreadyClosedException(CloseReason!);
                 }
             }
             else
             {
-                OnShutdown();
+                OnShutdown(reason);
                 _session0.SetSessionClosing(false);
 
                 try
                 {
-                    // Try to send connection.close
-                    // Wait for CloseOk in the MainLoop
-                    _session0.Transmit(new OutgoingCommand(new Impl.ConnectionClose(reason.ReplyCode, reason.ReplyText, 0, 0)));
+                    // Try to send connection.close wait for CloseOk in the MainLoop
+                    _session0.Transmit(new OutgoingCommand(new ConnectionClose(reason.ReplyCode, reason.ReplyText, 0, 0)));
                 }
                 catch (AlreadyClosedException)
                 {
@@ -318,55 +313,24 @@ namespace RabbitMQ.Client.Framing.Impl
             }
         }
 
-        private void ClosingLoop()
+        internal void InternalClose(ShutdownEventArgs reason)
         {
-            try
+            if (!SetCloseReason(reason))
             {
-                _frameHandler.ReadTimeout = TimeSpan.Zero;
-                // Wait for response/socket closure or timeout
-                while (!_closed)
+                if (_closed)
                 {
-                    MainLoopIteration();
+                    ThrowAlreadyClosedException(CloseReason!);
                 }
+                // We are quiescing, but still allow for server-close
             }
-            catch (ObjectDisposedException ode)
-            {
-                if (!_closed)
-                {
-                    LogCloseError("Connection didn't close cleanly", ode);
-                }
-            }
-            catch (EndOfStreamException eose)
-            {
-                if (_model0.CloseReason is null)
-                {
-                    LogCloseError("Connection didn't close cleanly. Socket closed unexpectedly", eose);
-                }
-            }
-            catch (IOException ioe)
-            {
-                LogCloseError("Connection didn't close cleanly. Socket closed unexpectedly", ioe);
-            }
-            catch (Exception e)
-            {
-                LogCloseError("Unexpected exception while closing: ", e);
-            }
-        }
 
-        internal ISession CreateSession()
-        {
-            return _sessionManager.Create();
-        }
-
-        internal void EnsureIsOpen()
-        {
-            if (!IsOpen)
-            {
-                throw new AlreadyClosedException(CloseReason);
-            }
+            OnShutdown(reason);
+            _session0.SetSessionClosing(true);
+            TerminateMainloop();
         }
 
         // Only call at the end of the Mainloop or HeartbeatLoop
+
         private void FinishClose()
         {
             _closed = true;
@@ -377,56 +341,16 @@ namespace RabbitMQ.Client.Framing.Impl
             _model0.FinishClose();
         }
 
-        private void HandleMainLoopException(ShutdownEventArgs reason)
+        ///<summary>Broadcasts notification of the final shutdown of the connection.</summary>
+        private void OnShutdown(ShutdownEventArgs reason)
         {
-            if (!SetCloseReason(reason))
-            {
-                LogCloseError("Unexpected Main Loop Exception while closing: " + reason, new Exception(reason.ToString()));
-                return;
-            }
-
-            OnShutdown();
-            LogCloseError($"Unexpected connection closure: {reason}", new Exception(reason.ToString()));
+            ThrowIfDisposed();
+            _connectionShutdownWrapper.Invoke(this, reason);
         }
 
-        private bool HardProtocolExceptionHandler(HardProtocolException hpe)
+        private bool SetCloseReason(ShutdownEventArgs reason)
         {
-            if (SetCloseReason(hpe.ShutdownReason))
-            {
-                OnShutdown();
-                _session0.SetSessionClosing(false);
-                try
-                {
-                    _session0.Transmit(new OutgoingCommand(new ConnectionClose(hpe.ShutdownReason.ReplyCode, hpe.ShutdownReason.ReplyText, 0, 0)));
-                    return true;
-                }
-                catch (IOException ioe)
-                {
-                    LogCloseError("Broker closed socket unexpectedly", ioe);
-                }
-            }
-            else
-            {
-                LogCloseError("Hard Protocol Exception occured while closing the connection", hpe);
-            }
-
-            return false;
-        }
-
-        internal void InternalClose(ShutdownEventArgs reason)
-        {
-            if (!SetCloseReason(reason))
-            {
-                if (_closed)
-                {
-                    throw new AlreadyClosedException(CloseReason);
-                }
-                // We are quiescing, but still allow for server-close
-            }
-
-            OnShutdown();
-            _session0.SetSessionClosing(true);
-            TerminateMainloop();
+            return System.Threading.Interlocked.CompareExchange(ref _closeReason, reason, null) is null;
         }
 
         private void LogCloseError(string error, Exception ex)
@@ -442,366 +366,14 @@ namespace RabbitMQ.Client.Framing.Impl
             }
         }
 
-        private void MainLoop()
-        {
-            bool shutdownCleanly = false;
-            try
-            {
-                while (_running)
-                {
-                    try
-                    {
-                        MainLoopIteration();
-                    }
-                    catch (SoftProtocolException spe)
-                    {
-                        QuiesceChannel(spe);
-                    }
-                }
-                shutdownCleanly = true;
-            }
-            catch (EndOfStreamException eose)
-            {
-                // Possible heartbeat exception
-                HandleMainLoopException(new ShutdownEventArgs(
-                    ShutdownInitiator.Library,
-                    0,
-                    "End of stream",
-                    eose));
-            }
-            catch (HardProtocolException hpe)
-            {
-                shutdownCleanly = HardProtocolExceptionHandler(hpe);
-            }
-            catch (Exception ex)
-            {
-                HandleMainLoopException(new ShutdownEventArgs(ShutdownInitiator.Library,
-                    Constants.InternalError,
-                    "Unexpected Exception",
-                    ex));
-            }
-
-            // If allowed for clean shutdown, run main loop until the
-            // connection closes.
-            if (shutdownCleanly)
-            {
-#pragma warning disable 0168
-                try
-                {
-                    ClosingLoop();
-                }
-                catch (SocketException)
-                {
-                    // means that socket was closed when frame handler
-                    // attempted to use it. Since we are shutting down,
-                    // ignore it.
-                }
-#pragma warning restore 0168
-            }
-
-            FinishClose();
-        }
-
-        private void MainLoopIteration()
-        {
-            InboundFrame frame = _frameHandler.ReadFrame();
-            NotifyHeartbeatListener();
-
-            bool shallReturn = true;
-            // We have received an actual frame.
-            if (frame.Type == FrameType.FrameHeartbeat)
-            {
-                // Ignore it: we've already just reset the heartbeat
-            }
-            else if (frame.Channel == 0)
-            {
-                // In theory, we could get non-connection.close-ok
-                // frames here while we're quiescing (m_closeReason !=
-                // null). In practice, there's a limited number of
-                // things the server can ask of us on channel 0 -
-                // essentially, just connection.close. That, combined
-                // with the restrictions on pipelining, mean that
-                // we're OK here to handle channel 0 traffic in a
-                // quiescing situation, even though technically we
-                // should be ignoring everything except
-                // connection.close-ok.
-                shallReturn = _session0.HandleFrame(in frame);
-            }
-            else
-            {
-                // If we're still m_running, but have a m_closeReason,
-                // then we must be quiescing, which means any inbound
-                // frames for non-zero channels (and any inbound
-                // commands on channel zero that aren't
-                // Connection.CloseOk) must be discarded.
-                if (_closeReason is null)
-                {
-                    // No close reason, not quiescing the
-                    // connection. Handle the frame. (Of course, the
-                    // Session itself may be quiescing this particular
-                    // channel, but that's none of our concern.)
-                    shallReturn = _sessionManager.Lookup(frame.Channel).HandleFrame(in frame);
-                }
-            }
-
-            if (shallReturn)
-            {
-                frame.ReturnPayload();
-            }
-        }
-
-        private void NotifyHeartbeatListener()
-        {
-            _heartbeatDetected = true;
-        }
-
-        private void NotifyReceivedCloseOk()
-        {
-            TerminateMainloop();
-            _closed = true;
-        }
-
         internal void OnCallbackException(CallbackExceptionEventArgs args)
         {
             _callbackExceptionWrapper.Invoke(this, args);
         }
 
-        ///<summary>Broadcasts notification of the final shutdown of the connection.</summary>
-        private void OnShutdown()
-        {
-            ThrowIfDisposed();
-            _connectionShutdownWrapper.Invoke(this, CloseReason);
-        }
-
-        private void Open()
-        {
-            StartAndTune();
-            _model0.ConnectionOpen(_factory.VirtualHost);
-        }
-
-        ///<summary>
-        /// Sets the channel named in the SoftProtocolException into
-        /// "quiescing mode", where we issue a channel.close and
-        /// ignore everything except for subsequent channel.close
-        /// messages and the channel.close-ok reply that should
-        /// eventually arrive.
-        ///</summary>
-        ///<remarks>
-        ///<para>
-        /// Since a well-behaved peer will not wait indefinitely before
-        /// issuing the close-ok, we don't bother with a timeout here;
-        /// compare this to the case of a connection.close-ok, where a
-        /// timeout is necessary.
-        ///</para>
-        ///<para>
-        /// We need to send the close method and politely wait for a
-        /// reply before marking the channel as available for reuse.
-        ///</para>
-        ///<para>
-        /// As soon as SoftProtocolException is detected, we should stop
-        /// servicing ordinary application work, and should concentrate
-        /// on bringing down the channel as quickly and gracefully as
-        /// possible. The way this is done, as per the close-protocol,
-        /// is to signal closure up the stack *before* sending the
-        /// channel.close, by invoking ISession.Close. Once the upper
-        /// layers have been signalled, we are free to do what we need
-        /// to do to clean up and shut down the channel.
-        ///</para>
-        ///</remarks>
-        private void QuiesceChannel(SoftProtocolException pe)
-        {
-            // Construct the QuiescingSession that we'll use during
-            // the quiesce process.
-
-            ISession newSession = new QuiescingSession(this,
-                (ushort)pe.Channel,
-                pe.ShutdownReason);
-
-            // Here we detach the session from the connection. It's
-            // still alive: it just won't receive any further frames
-            // from the mainloop (once we return to the mainloop, of
-            // course). Instead, those frames will be directed at the
-            // new QuiescingSession.
-            ISession oldSession = _sessionManager.Swap(pe.Channel, newSession);
-
-            // Now we have all the information we need, and the event
-            // flow of the *lower* layers is set up properly for
-            // shutdown. Signal channel closure *up* the stack, toward
-            // the model and application.
-            oldSession.Close(pe.ShutdownReason);
-
-            // The upper layers have been signalled. Now we can tell
-            // our peer. The peer will respond through the lower
-            // layers - specifically, through the QuiescingSession we
-            // installed above.
-            newSession.Transmit(new OutgoingCommand(new Impl.ChannelClose(pe.ReplyCode, pe.Message, 0, 0)));
-        }
-
-        private bool SetCloseReason(ShutdownEventArgs reason)
-        {
-            return Interlocked.CompareExchange(ref _closeReason, reason, null) is null;
-        }
-
-        private void MaybeStartHeartbeatTimers()
-        {
-            if (Heartbeat != TimeSpan.Zero)
-            {
-                _heartbeatWriteTimer ??= new Timer(HeartbeatWriteTimerCallback, null, 200, Timeout.Infinite);
-                _heartbeatReadTimer ??= new Timer(HeartbeatReadTimerCallback, null, 300, Timeout.Infinite);
-            }
-        }
-
-        private void StartMainLoop()
-        {
-            _mainLoopTask = Task.Factory.StartNew(MainLoop, TaskCreationOptions.LongRunning);
-        }
-
-        private void HeartbeatReadTimerCallback(object state)
-        {
-            if (_heartbeatReadTimer is null)
-            {
-                return;
-            }
-
-            bool shouldTerminate = false;
-
-            try
-            {
-                if (!_closed)
-                {
-                    if (_heartbeatDetected)
-                    {
-                        _heartbeatDetected = false;
-                        _missedHeartbeats = 0;
-                    }
-                    else
-                    {
-                        _missedHeartbeats++;
-                    }
-
-                    // We check against 8 = 2 * 4 because we need to wait for at
-                    // least two complete heartbeat setting intervals before
-                    // complaining, and we've set the socket timeout to a quarter
-                    // of the heartbeat setting in setHeartbeat above.
-                    if (_missedHeartbeats > 2 * 4)
-                    {
-                        var eose = new EndOfStreamException($"Heartbeat missing with heartbeat == {_heartbeat} seconds");
-                        LogCloseError(eose.Message, eose);
-                        HandleMainLoopException(new ShutdownEventArgs(ShutdownInitiator.Library, 0, "End of stream", eose));
-                        shouldTerminate = true;
-                    }
-                }
-
-                if (shouldTerminate)
-                {
-                    TerminateMainloop();
-                    FinishClose();
-                }
-                else
-                {
-                    _heartbeatReadTimer?.Change((int)Heartbeat.TotalMilliseconds, Timeout.Infinite);
-                }
-            }
-            catch (ObjectDisposedException)
-            {
-                // timer is already disposed,
-                // e.g. due to shutdown
-            }
-            catch (NullReferenceException)
-            {
-                // timer has already been disposed from a different thread after null check
-                // this event should be rare
-            }
-        }
-
-        private void HeartbeatWriteTimerCallback(object state)
-        {
-            if (_heartbeatWriteTimer is null)
-            {
-                return;
-            }
-
-            try
-            {
-                if (!_closed)
-                {
-                    Write(Client.Impl.Framing.Heartbeat.GetHeartbeatFrame());
-                    _heartbeatWriteTimer?.Change((int)_heartbeatTimeSpan.TotalMilliseconds, Timeout.Infinite);
-                }
-            }
-            catch (ObjectDisposedException)
-            {
-                // timer is already disposed,
-                // e.g. due to shutdown
-            }
-            catch (Exception)
-            {
-                // ignore, let the read callback detect
-                // peer unavailability. See rabbitmq/rabbitmq-dotnet-client#638 for details.
-            }
-        }
-
-        private void MaybeStopHeartbeatTimers()
-        {
-            NotifyHeartbeatListener();
-            _heartbeatReadTimer?.Dispose();
-            _heartbeatWriteTimer?.Dispose();
-        }
-
-        ///<remarks>
-        /// May be called more than once. Should therefore be idempotent.
-        ///</remarks>
-        private void TerminateMainloop()
-        {
-            MaybeStopHeartbeatTimers();
-            _running = false;
-        }
-
-        public override string ToString()
-        {
-            return $"Connection({_id},{Endpoint})";
-        }
-
         internal void Write(ReadOnlyMemory<byte> memory)
         {
             _frameHandler.Write(memory);
-        }
-
-        public void UpdateSecret(string newSecret, string reason)
-        {
-            _model0.UpdateSecret(newSecret, reason);
-        }
-
-        ///<summary>API-side invocation of connection.close with timeout.</summary>
-        public void Close(ushort reasonCode, string reasonText, TimeSpan timeout, bool abort)
-        {
-            Close(new ShutdownEventArgs(ShutdownInitiator.Application, reasonCode, reasonText), abort, timeout);
-        }
-
-        public IModel CreateModel()
-        {
-            EnsureIsOpen();
-            ISession session = CreateSession();
-            var model = (ModelBase)Protocol.CreateModel(_factory, session);
-            model.ContinuationTimeout = _factory.ContinuationTimeout;
-            model._Private_ChannelOpen();
-            return model;
-        }
-
-        internal void HandleConnectionBlocked(string reason)
-        {
-            if (!_connectionBlockedWrapper.IsEmpty)
-            {
-                _connectionBlockedWrapper.Invoke(this, new ConnectionBlockedEventArgs(reason));
-            }
-        }
-
-        internal void HandleConnectionUnblocked()
-        {
-            if (!_connectionUnblockedWrapper.IsEmpty)
-            {
-                _connectionUnblockedWrapper.Invoke(this, EventArgs.Empty);
-            }
         }
 
         public void Dispose()
@@ -826,127 +398,28 @@ namespace RabbitMQ.Client.Framing.Impl
             }
         }
 
-        private void StartAndTune()
-        {
-            var connectionStartCell = new BlockingCell<ConnectionStartDetails>();
-            _model0.m_connectionStartCell = connectionStartCell;
-            _model0.HandshakeContinuationTimeout = _factory.HandshakeContinuationTimeout;
-            _frameHandler.ReadTimeout = _factory.HandshakeContinuationTimeout;
-            _frameHandler.SendHeader();
-
-            ConnectionStartDetails connectionStart = connectionStartCell.WaitForValue();
-
-            if (connectionStart is null)
-            {
-                throw new IOException("connection.start was never received, likely due to a network timeout");
-            }
-
-            ServerProperties = connectionStart.m_serverProperties;
-
-            var serverVersion = new AmqpVersion(connectionStart.m_versionMajor, connectionStart.m_versionMinor);
-            if (!serverVersion.Equals(Protocol.Version))
-            {
-                TerminateMainloop();
-                FinishClose();
-                throw new ProtocolVersionMismatchException(Protocol.MajorVersion, Protocol.MinorVersion, serverVersion.Major, serverVersion.Minor);
-            }
-
-            ClientProperties = new Dictionary<string, object>(_factory.ClientProperties)
-            {
-                ["capabilities"] = Protocol.Capabilities,
-                ["connection_name"] = ClientProvidedName
-            };
-
-            // FIXME: parse out locales properly!
-            ConnectionTuneDetails connectionTune = default;
-            bool tuned = false;
-            try
-            {
-                string mechanismsString = Encoding.UTF8.GetString(connectionStart.m_mechanisms, 0, connectionStart.m_mechanisms.Length);
-                string[] mechanisms = mechanismsString.Split(' ');
-                IAuthMechanismFactory mechanismFactory = _factory.AuthMechanismFactory(mechanisms);
-                if (mechanismFactory is null)
-                {
-                    throw new IOException($"No compatible authentication mechanism found - server offered [{mechanismsString}]");
-                }
-                IAuthMechanism mechanism = mechanismFactory.GetInstance();
-                byte[] challenge = null;
-                do
-                {
-                    byte[] response = mechanism.handleChallenge(challenge, _factory);
-                    ConnectionSecureOrTune res;
-                    if (challenge is null)
-                    {
-                        res = _model0.ConnectionStartOk(ClientProperties,
-                            mechanismFactory.Name,
-                            response,
-                            "en_US");
-                    }
-                    else
-                    {
-                        res = _model0.ConnectionSecureOk(response);
-                    }
-
-                    if (res.m_challenge is null)
-                    {
-                        connectionTune = res.m_tuneDetails;
-                        tuned = true;
-                    }
-                    else
-                    {
-                        challenge = res.m_challenge;
-                    }
-                }
-                while (!tuned);
-            }
-            catch (OperationInterruptedException e)
-            {
-                if (e.ShutdownReason != null && e.ShutdownReason.ReplyCode == Constants.AccessRefused)
-                {
-                    throw new AuthenticationFailureException(e.ShutdownReason.ReplyText);
-                }
-                throw new PossibleAuthenticationFailureException(
-                    "Possibly caused by authentication failure", e);
-            }
-
-            ushort channelMax = (ushort)NegotiatedMaxValue(_factory.RequestedChannelMax, connectionTune.m_channelMax);
-            _sessionManager = new SessionManager(this, channelMax);
-
-            uint frameMax = NegotiatedMaxValue(_factory.RequestedFrameMax, connectionTune.m_frameMax);
-            FrameMax = frameMax;
-
-            TimeSpan requestedHeartbeat = _factory.RequestedHeartbeat;
-            uint heartbeatInSeconds = NegotiatedMaxValue((uint)requestedHeartbeat.TotalSeconds, (uint)connectionTune.m_heartbeatInSeconds);
-            Heartbeat = TimeSpan.FromSeconds(heartbeatInSeconds);
-
-            _model0.ConnectionTuneOk(channelMax, frameMax, (ushort)Heartbeat.TotalSeconds);
-
-            // now we can start heartbeat timers
-            MaybeStartHeartbeatTimers();
-        }
-
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         private void ThrowIfDisposed()
         {
             if (_disposed)
             {
-                throw new ObjectDisposedException(GetType().FullName);
+                ThrowObjectDisposedException();
+            }
+
+            static void ThrowObjectDisposedException()
+            {
+                throw new ObjectDisposedException(typeof(Connection).FullName);
             }
         }
 
-        private static uint NegotiatedMaxValue(uint clientValue, uint serverValue)
+        public override string ToString()
         {
-            return (clientValue == 0 || serverValue == 0) ?
-                Math.Max(clientValue, serverValue) :
-                Math.Min(clientValue, serverValue);
+            return $"Connection({_id},{Endpoint})";
         }
 
-        internal void TakeOver(Connection other)
+        private static void ThrowAlreadyClosedException(ShutdownEventArgs closeReason)
         {
-            _callbackExceptionWrapper.Takeover(other._callbackExceptionWrapper);
-            _connectionBlockedWrapper.Takeover(other._connectionBlockedWrapper);
-            _connectionUnblockedWrapper.Takeover(other._connectionUnblockedWrapper);
-            _connectionShutdownWrapper.Takeover(other._connectionShutdownWrapper);
+            throw new AlreadyClosedException(closeReason);
         }
     }
 }
